MVP of native IO support using DDRIO interface for ksys573
This is fully working but still not optimal regarding performance.
Since the polling is done as part of some main processing loop,
this introduces significant latency and slows things down on
rather low spec setups that the game can't run at 100% speed
anymore (but close it to ~90%).

Therefore, this needs some more research on how the current IO
code is called and coming up with either optimisations or a new
solution to solve the performance issue.