ksys573: Support system buttons with DDRIO
Use a different approach to hook into the inputs of single bits
of the port mappings. Using a full custom handler on the whole
register doesn't work with all the different bits that have
to communicate with other hardware parts.

Therefore, use the read line custom port definition to read and
modify a single bit only. Define another port just for system
inputs from joystick device and merge them back with the
DDRIO inputs in the custom hooks.